(maint) Add Partial Parsing
This commit is the beginnings of adding parsing of the stats file to
obtain data needed for calculations.

This commit also adds graphing on the project action page for the two
graph calculations we are interested in. For now it is graphing fake
data, but once all parsing is complete it will be graphing those
calculated points.